Undo erroneous type change

+++ pyuvdata/parameter.py
@@ -326,14 +326,14 @@ class UVParameter(object):
         """
         if self.form == "str":
             return self.form
-        elif isinstance(self.form, (int, np.int_)):
+        elif isinstance(self.form, (int, np.integer)):
             # Fixed shape, just return the form
             return (self.form,)
         else:
             # Given by other attributes, look up values
             eshape = ()
             for p in self.form:
-                if isinstance(p, (int, np.int_)):
+                if isinstance(p, (int, np.integer)):
                     eshape = eshape + (p,)
                 else:
                     val = getattr(uvbase, p)
